+++ packages/skills/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @agent-native/skills
 
+## 0.2.618
+
+### Patch Changes
+
+- Updated dependencies [9dd50a0]
+- Updated dependencies [f294ae3]
+  - @agent-native/core@0.161.18
+
+## 0.2.617
+
+### Patch Changes
+
+- Updated dependencies [34496d7]
+  - @agent-native/core@0.161.17
+
+## 0.2.616
+
+### Patch Changes
+
+- Updated dependencies [c940f4c]
+  - @agent-native/core@0.161.16
+
 ## 0.2.615
 
 ### Patch Changes